Fixed several issues :   - Fixed issue when game list > 25   - Fixed issue with ban button   - Added amd64 as a platform in GitHub actions CI script   - Refactored a bit games-roles module
