+++ .github/workflows/docker-image.yml
@@ -47,18 +47,15 @@ jobs:
     runs-on: ubuntu-latest
     needs: checkAndBuild
     env:
+      DOCKER_REGISTRY_HOST: docker-registry.pandore.ovh
       IMAGE_NAME: bunny-bot
-      DOCKER_TARGET_PLATFORM: linux/arm64
-      DOCKER_REGISTRY: docker-registry.pandore.ovh      
-      DOCKER_USERNAME: ${{ secrets.DOCKER_REGISTRY_USERNAME }}
-      DOCKER_PASSWORD: ${{ secrets.DOCKER_REGISTRY_PASSWORD }}
 
     steps:
     - name: Checkout repository
-      uses: actions/checkout@v2
+      uses: actions/checkout@v3
       
     - name: Set up Docker Buildx      
-      uses: crazy-max/ghaction-docker-buildx@v1
+      uses: crazy-max/ghaction-docker-buildx@v3
       with:
         version: latest
 
@@ -71,30 +68,10 @@ jobs:
         name: dist
         path: dist
 
-    - name: Prepare build (pull request)
-      if: success() && github.event_name == 'pull_request'
-      id: prepare-integration
+    - name: Prepare variables
+      id: prepare-variables
       run: |
-        IMAGE_ID=${{ secrets.DOCKER_REGISTRY_HOST }}/$IMAGE_NAME
-    
-        # Change all uppercase to lowercase
-        IMAGE_ID=$(echo $IMAGE_ID | tr '[A-Z]' '[a-z]')
-          
-        # Define version as "integration"
-        VERSION=integration
-
-        # Strip "v" prefix from tag name
-        [[ "${{ github.ref }}" == "refs/tags/"* ]] && VERSION=$(echo $VERSION | sed -e 's/^v//')
-
-        echo ::set-output name=docker_platform::${DOCKER_TARGET_PLATFORM}
-        echo ::set-output name=docker_image::$IMAGE_ID
-        echo ::set-output name=version::$VERSION
-    
-    - name: Prepare build (deployment)
-      if: success() && github.event_name != 'pull_request'
-      id: prepare-latest
-      run: |
-        IMAGE_ID=${{ secrets.DOCKER_REGISTRY_HOST }}/$IMAGE_NAME
+        IMAGE_ID=${{ env.DOCKER_REGISTRY_HOST }}/$IMAGE_NAME
     
         # Change all uppercase to lowercase
         IMAGE_ID=$(echo $IMAGE_ID | tr '[A-Z]' '[a-z]')
@@ -108,23 +85,18 @@ jobs:
         # Use Docker `latest` tag convention
         [ "$VERSION" == "master" ] && VERSION=latest
 
-        echo ::set-output name=docker_platform::${DOCKER_TARGET_PLATFORM}
-        echo ::set-output name=docker_image::$IMAGE_ID
-        echo ::set-output name=version::$VERSION
+        # If integration
+        [ "${{ github.event_name }}" == "pull_request" ] && VERSION=integration
+
+        echo ::set-output name=image::$IMAGE_ID
+        echo ::set-output name=tag::$VERSION
 
     - name: Login to Docker Registry
       if: success()
       run: |        
-        echo "${DOCKER_PASSWORD}" | docker login ${DOCKER_REGISTRY} --username "${DOCKER_USERNAME}" --password-stdin
+        echo "${{ secrets.DOCKER_REGISTRY_PASSWORD }}" | docker login ${{ env.DOCKER_REGISTRY_HOST }} --username "${{ secrets.DOCKER_REGISTRY_USERNAME }}" --password-stdin
 
     - name: Run Buildx (Build and push to integration)
-      if: success() && github.event_name == 'pull_request'
+      if: success()
       run: |        
-        docker buildx build --platform ${{ steps.prepare-integration.outputs.docker_platform }} --tag ${{ steps.prepare-integration.outputs.docker_image }}:${{ steps.prepare-integration.outputs.version }} --output type=image --push .
-
-    - name: Run Buildx (Build and push to production)
-      if: success() && github.event_name != 'pull_request'
-      run: |
-        docker buildx build --platform ${{ steps.prepare-latest.outputs.docker_platform }} --tag ${{ steps.prepare-latest.outputs.docker_image }}:${{ steps.prepare-latest.outputs.version }} --output type=image --push .
-
-    
+        docker buildx build --platform linux/amd64,linux/arm64 --tag ${{ steps.prepare-variables.outputs.image }}:${{ steps.prepare-variables.outputs.tag }} --output "type=image,push=true" .
