+++ src/modules/games-roles/helpers/message.helper.ts
@@ -4,124 +4,6 @@ import { Game, IGame } from "../models/game.model";
 
 export class MessageHelper {
 
-  /**
-   * Sends modal for game creation
-   * 
-   * @param interaction Button interaction pressed to send the modal
-   */
-  public static async sendCreateGameModal(interaction: ButtonInteraction): Promise<void> {
-    
-    // Create the modal
-    const modal = new Modal()
-      .setCustomId('games-roles-create-game')
-      .setTitle('Add a new game in my list');
-
-    // Add components to modal
-    const gameInput = new TextInputComponent()
-      .setCustomId('game-name')
-      .setLabel("What's the game name?")
-      .setStyle('SHORT');
-
-    modal.addComponents(
-      new MessageActionRow<ModalActionRowComponent>().addComponents(gameInput)
-    );
-  
-    // Show the modal to the user
-    await interaction.showModal(modal);
-  }
-
-  /**
-   * Sends select menu to choose a game to activate
-   * 
-   * @param interaction Button interaction which should be replied
-   */
-  public static async sendActivateSelectMenu(interaction: ButtonInteraction): Promise<void> {
-
-    await interaction.deferUpdate();
-
-    const messageActionRow = new MessageActionRow();
-    const selectMenu = new MessageSelectMenu();
-    selectMenu.setCustomId('games-roles-activate');
-    selectMenu.setPlaceholder('Choose a game');
-
-    const games = await Game.getGamesThatCanBeActivatedForGuild(interaction.guildId);
-    const options: MessageSelectOptionData[] = games.map((game: IGame) => { return { label: game.gameName, value: game.applicationId }; });
-
-    selectMenu.setOptions(options);
-    messageActionRow.addComponents(selectMenu);
-    
-    await interaction.editReply({ content: 'Choose a game to activate', embeds: [], components: [ messageActionRow ] });
-  }
-
-  /**
-   * Sends select menu to choose a game to deactivate
-   * 
-   * @param interaction Button interaction which should be replied
-   */
-   public static async sendDeactivateSelectMenu(interaction: ButtonInteraction): Promise<void> {
-
-    await interaction.deferUpdate();
-
-    const messageActionRow = new MessageActionRow();
-    const selectMenu = new MessageSelectMenu();
-    selectMenu.setCustomId('games-roles-deactivate');
-    selectMenu.setPlaceholder('Choose a game');
-
-    const games = await Game.getGamesThatCanBeDeactivatedForGuild(interaction.guildId);
-    const options: MessageSelectOptionData[] = games.map((game: IGame) => { return { label: game.gameName, value: game.applicationId }; });
-
-    selectMenu.setOptions(options);
-    messageActionRow.addComponents(selectMenu);
-    
-    await interaction.editReply({ content: 'Choose a game to deactivate', embeds: [], components: [ messageActionRow ] });
-  }
-
-  /**
-   * Sends select menu to choose a game to archive
-   * 
-   * @param interaction  Button interaction which should be replied
-   */
-  public static async sendArchiveSelectMenu(interaction: ButtonInteraction): Promise<void> {
-
-    await interaction.deferUpdate();
-
-    const messageActionRow = new MessageActionRow();
-    const selectMenu = new MessageSelectMenu();
-    selectMenu.setCustomId('games-roles-archive');
-    selectMenu.setPlaceholder('Choose a game to archive');
-
-    const games = await Game.getGamesWithChannelsForGuild(interaction.guildId);
-    const options: MessageSelectOptionData[] = games.map((game: IGame) => { return { label: game.gameName, value: game.applicationId }; });
-
-    selectMenu.setOptions(options);
-    messageActionRow.addComponents(selectMenu);
-    
-    await interaction.editReply({ content: 'Choose a game to archive', embeds: [], components: [ messageActionRow ] });
-  }
-
-  /**
-   * Sends select menu to choose a game to unarchive
-   * 
-   * @param interaction Button interaction which should be replied
-   */
-  public static async sendUnArchiveSelectMenu(interaction: ButtonInteraction): Promise<void> {
-
-    await interaction.deferUpdate();
-
-    const messageActionRow = new MessageActionRow();
-    const selectMenu = new MessageSelectMenu();
-    selectMenu.setCustomId('games-roles-unarchive');
-    selectMenu.setPlaceholder('Choose a game to unarchive');
-
-    const games = await Game.getArchivedGamesForGuild(interaction.guildId);
-    const options: MessageSelectOptionData[] = games.map((game: IGame) => { return { label: game.gameName, value: game.applicationId }; });
-
-    selectMenu.setOptions(options);
-    messageActionRow.addComponents(selectMenu);
-    
-    await interaction.editReply({ content: 'Choose a game to unarchive', embeds: [], components: [ messageActionRow ] });
-  }
-
   public static async sendSendMessageModal(interaction: ButtonInteraction): Promise<void> {
 
     // Create the modal
